Bug Fix Edition 2023 4
Mistakenly output double underlines when the title on the first page is only of one line. This bug has emerged, because when assigning the special auxiliary codes to output the title, the original mistaken codes ask ''the font setter of the content of title'' to be ''the font environment setter of the content line of title''.